MONDRIAN: Fix numerous issues found by 'findbugs'. 	Fix dialect: MySQL allows subquery in FROM clause from 4.0 onwards.
[git-p4: depot-paths = "//open/mondrian/": change = 9303]

+++ testsrc/main/mondrian/test/QueryRunner.java
@@ -215,7 +215,7 @@ public class QueryRunner extends Thread {
     };
 
     public QueryRunner(int id, int numSeconds, boolean useRandomQuery) {
-        mRunTime = numSeconds * 1000;
+        mRunTime = (long) numSeconds * 1000;
         mMyId = id;
         mRandomQueries = useRandomQuery;
     }
